% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Generates a map of the fatalities occurring in a given state in a given year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{An integer corresponding to a required state.
The function throws an error if no fatality information is available for
the requested state in the FARS data for the requested year.}

\item{year}{A year, provided as a string or an integer. If
the year is absent from the FARS data, this throws an error.}
}
\value{
NULL return value. The function plots a graph as a side-effect
  If no accidents were observed in the state/year, no map is plotted.
}
\description{
Provided a state (as an integer) and a year, this obtains all fatality data
for that state and plots out all the positions of the fatalities for that
state on a map.
}
\examples{
fars_map_state(10, 2013)

}
