% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Reads 'Fatality Analysis Reporting System' data from a file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{The path to a .csv file that contains the FARS
dataset.}
}
\value{
A tibble containing the FARS dataset.
}
\description{
Given a path to a file that contains the FARS dataset, this function will
read that dataset in and return a tibble containing the data. Throws an
error if the filepath does not exist.
}
\examples{
\dontrun{fars_read("some_file.csv")}

}
