% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Obtains the 'Fatality Analysis Reporting System' data for a set of years
from the whole set of FARS data.}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector of years, provided as a character
vector or as an integer vector. Provide the years in 4-digit format
to ensure the correct file can be found for that year. If any of the years
is absent from the FARS dataset, a NULL value is returned for that year.}
}
\value{
A list of tibble data-frames containing the FARS dataset for
  each year in the input set. The entry is NULL for any year that is absent
  from the FARS dataset.
}
\description{
Takes a vector of years, imports the data for each year from the FARS
dataset and returns a list of tibbles containing the data for each year.
}
\examples{
fars_read_years(c(2013, 2015))

}
