% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarises the FARS data for a set of years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector of years, provided as integers or 
strings. Only those entries that are present in the FARS data provide any
valid returned data.}
}
\value{
A tibble containing the count of fatalities in the FARS data
  for each month and year. If all of the years are absent from the FARS data
  a NULL value is returned.
}
\description{
Given a vector of years, this function will obtain all FARS data for each
year and combine it into a summarised dataset. The count of fatailities in 
each month in the selected years is returned. No data is returned for years
that are absent from the available data.
}
\examples{
fars_summarize_years(1984:1987)

}
