% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Converts a year into the filename that contains FARS data for that year.}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A vector of years, provided as integers (1984)
or strings ("1984", "2001"). Do not provide R 'Date' objects or shortened
suffixes (eg 84) for a year.}
}
\value{
A vector of filenames.
}
\description{
The FARS data is stored in a range of .bz2 files with a consistent format.
This function takes a year (as a string or an integer, like '1984') and
returns the .bz2 file corresponding to that year
}
